Add an automatic index to the EntityConcernedTypesAllowed element.
If, for example, a KVK and RSIN attribute is specified in the
'entity_concerned_types_allowed' configuration list, it allows the
user to login using either KVK or RSIN.

With this index we automatically create an index set for each attribute
specified. In the future we might want to add the possibility to have
multiple attributes in the same identifier set.