:loud_sound: Replace fatal error with logging
There appear to be ~1 incident/day caused by this IP address check,
and from the looks of it they all seems like genuine, non-malicious
usage.

In office spaces where VPNs and network changes (which result in
different IP addresses) are common, this check leads to false positives
and we've assessed the likelihood of actual attacks to be low.

* Session cookies are typically flagged as HttpOnly, so XSS to steal
  credentials/cookies is not a viable exploit path
* If an attacker can intercept the session cookie via MITM over plain
  HTTP, the DigiD conditions are not met and you really, really should
  not run anything over plain HTTP.
* If an attacker can MITM an HTTPS connection, you have far bigger
  issues, since then interecepting/modifying the actual HTTP traffic is
  much more interesting than hijacking a DigiD session.
* Initial HTTP connections (first-ever) are possible, even when using
  HSTS (which you should deploy), but can be mitigated by making sure
  your URLs are added to the preload lists vendored in browsers.