:card_file_box: Add LOA mapping configuration field
Different brokers (Anoigo, Signicat) seem to return the LOA from the
used authentication mean in a non-standard way, so our configuration
needs to be able to translate their values into standardized values.

Brokers that don't return a value at all can be configured to specify
a default value to apply.

Source values (from the broker) can be string or number, destination
values are enum-based depending on whether it's DigiD or eHerkenning.