:sparkles: [#75] Offer metadata as file for download
Using Content-Disposition attachment offers the file for download to
the client rather than viewing it inline.

The actual reason the dislaying was botched was because of the content
security policy which blocked styles (not sure how that works).
However, if that were to be addressed, it leaves the certificate data
newlines being displayed as spaces, still producing content that is
not ready to copy-and-paste, so downloading a file is the best option.