:card_fiel_box: Implement data migration to decrypt private keys
Using the configured private key, decrypt the certificate's private
key and 'replace' the file to point to the decrypted private key
instead.

Security wise this makes little difference - if you had access to the
DB, you could already grab the passphrase to decrypt it yourself.

The private media storage and database access are expected to be
properly secured by your infra/ops team.