:fire: Delete metadata generation management commands
As these will need to be updated for the multi-certificate feature to
be supported, it was checked if they are used in the first place.
DevOps says it's not used, so removing them is far easier.

The metadata itself is still accessible through URLs/views and the
admin interface from each configuration page, it's just the CLI
integration that is removed.