:card_file_box: [#74] Add models and admin integration for digid/eh certificates
Added a proxy model so that we can expose/introspect the valid_from of
a given certificate, which will be needed to select the appropriate
certificate for authentication requests.

The model to link certificate and config deliberately uses a simple
constant - a foreign key or m2m doesn't make much sense because we're
dealing with singleton/django-solo models of which there will only
ever be a single instance. A flatter table layout is therefore possible.

Next steps will be adding some tests for the admin and a data migration
to copy the single certificate FK field over to the new structure.