:building_construction: [#74] Flip around how we acquire the signing key and certificate
The SAML client will still grab the configuration from our config dicts,
however the config dicts are now populated from the candidate selection
helper (to be implemented).

This allows us to add tests for edge cases to the certificate selection,
and de-duplicates the logic for certificate selection.

We are limited by python3-saml which only has hooks for the current and
new certificate, and it doesn't support an arbitrary amount of certificates.

Subclassing *could* be an option, but then we're forced to copy and
modify a lot of code that we actually don't care about and is too
risky to forget to update (and the same considerations apply
with making the changes in our fork).

Given the use case that we want to prepare for certificate
replacements, there doesn't seem to be a pressing urgency to
support more than 2 certificates in the metadata.