:sparkles: [#74] Ensure that the next certificate is included in the metata
When a next certificate to be used for signing is available/known,
ensure that it is passed along to the metadata generation. That way,
the new metadata can be exchanged, informing the identity provider
of our future certificate, allowing us to seamlessy transition
once the 'current' certificate expires.

python3-saml supports this through the x509certNew settings
dict property.