:bug: [open-formulieren/open-forms#4785] Remove the NameIDFormat element
While having this element present passes XSD validation against the
SAML 2.0 metadata schema, this is not accepted by brokers anymore
because of the line in the AS1.24a specification saying that unlisted
elements must not be included in the metadata.

I've opted to drop this key/element in the eHerkenning SAML client
implementation rather than the base class because I don't know if
removing it entirely will cause the DigiD metadata to break. It would
probably be wise to *not* share a common base class anymore for DigiD
and eHerkenning as it proves to be quite a maintenance nightmare.

Documentation at the time of writing: https://afsprakenstelsel.etoegang.nl/Startpagina/v3/dv-metadata-for-hm