:test_tube: [open-formulieren/open-forms#5136] Add regression test for old cert being used
If a new certificate has been configured and the service catalogue is
being generated again, the expectation is that the next certificate
is included in the metadata rather than the current, because the
latter will (likely) soon expire.

Note that this assumption breaks if users prepare the next
certificate way ahead of time (e.g. it's ready one year
before the current certificate expires), but this seems
mostly a theoretical case since certificates appear to
be issued and then taken into production in a matter of
hours or days.