:fire: [#96] Remove Digid/eHerkenning OIDC
As part of The Great OIDC Refactor, we realised that the only 2 Maykin projects using OIDC digid/ehekrenning were Open Forms and Open Inwoner. They both override the 'generic OIDC' code that was in this library. So, it was decided to remove this code and implement the plugins with the project specific logic directly in the projects.